Step 14 ex 5: ask user for filename for save and load

+++ directory.rb
@@ -16,12 +16,18 @@ def input_students
 end
 
 def load_students(filename = "students.csv")
+  if !File.exists?(filename)
+    puts "Error: File #{filename} does not exist"
+    return
+  end
+  puts "Loading the students..."
   file = File.open(filename, "r")
   file.readlines.each do |line|
     name, cohort = line.chomp.split(",")
     push_student(name, cohort)
   end
   file.close
+  puts "Students loaded"
 end
 
 def push_student(name, cohort)
@@ -41,6 +47,11 @@ def try_load_students
   end
 end
 
+def ask_filename
+  puts "Please enter filename (default students.csv)"
+  STDIN.gets.chomp
+end
+
 def interactive_menu
   loop do
     print_menu
@@ -51,8 +62,8 @@ end
 def print_menu
   puts "1. Input the students"
   puts "2. Show the students"
-  puts "3. Save the list to students.csv"
-  puts "4. Load the list from students.csv"
+  puts "3. Save the list to file"
+  puts "4. Load the list from file"
   puts "9. Exit" # 9 because we'll be adding more items
 end
 
@@ -69,12 +80,11 @@ def process(selection)
     when "2"
       show_students
     when "3"
-      save_students
-      puts "Students saved"
+      filename = ask_filename
+      filename.empty? ? save_students : save_students(filename)
     when "4"
-      puts "Loading the students..."
-      load_students
-      puts "Students loaded"
+      filename = ask_filename
+      filename.empty? ? load_students : load_students(filename)
     when "9"
       exit # this will cause the program to terminate
     else
@@ -97,9 +107,9 @@ def print_footer
   puts "Overall, we have #{@students.count} great students"
 end
 
-def save_students
+def save_students(filename = "students.csv")
   # open the file for writing
-  file = File.open("students.csv", "w")
+  file = File.open(filename, "w")
   # iterate over the array of students
   @students.each do |student|
     student_data = [student[:name], student[:cohort]]
@@ -107,6 +117,7 @@ def save_students
     file.puts csv_line
   end
   file.close
+  puts "Students saved"
 end
 
 try_load_students
